add oop code

+++ myproject-example/src/main/java/com/tangshengbo/design/oop/calc/Goods.java
@@ -5,5 +5,5 @@ package com.tangshengbo.design.oop.calc;
  */
 public interface Goods {
 
-    int getPrice();
+    float getPrice();
 }
